breakdown PopulationPrevalence into multiple components
